+++ immutable_views/_list_view.py
@@ -20,7 +20,7 @@ try:
     from collections.abc import Sequence
 except ImportError:
     # Python 2
-    from collections import Sequence
+    from collections import Sequence  # pylint: disable=deprecated-class
 
 __all__ = ['ListView']